% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brweather_read.R
\name{brweather_read}
\alias{brweather_read}
\title{Title}
\usage{
brweather_read(longitude, latitude, variables, time = NULL, path = ".")
}
\arguments{
\item{longitude}{longitude}

\item{latitude}{latitude}

\item{variables}{var}

\item{time}{time}

\item{path}{path}
}
\value{
Data.frame
}
\description{
Title
}
\examples{
wd <- brweather_read(
  longitude = -50.5995,
  latitude = -27.2863,
  variables = c("pr", "ETo"),
  time = c("1968-01-01", "1983-12-31"),
  path = "C:/Users/joaob/OneDrive - UFSC/clima_Xavier"
)
}
